requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
bitsandbytes==0.44.1
blinker==1.8.2
boto3==1.35.47
botocore==1.35.47
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colpali_engine==0.2.2
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.2
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docutils==0.21.2
durationpy==0.7
einops==0.8.0
elastic-transport==8.15.0
elasticsearch==8.15.1
Elixir==0.7.1
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.115.3
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
GPUtil==1.4.0
gradio==5.4.0
gradio_client==1.4.2
greenlet==3.1.1
gritql==0.1.5
groq==0.11.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imagesize==1.4.1
importlib_metadata==8.4.0
importlib_resources==6.4.5
indic_nlp_library_IT2 @ git+https://github.com/VarunGumma/indic_nlp_library@601521e05ed0ed8f2165ac317a47d186e25b6f0d
-e git+https://github.com/VarunGumma/IndicTransToolkit.git@0c607654e8b6cce897a32039df79018a1daff4e2#egg=IndicTransToolkit
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.4
langchain-chroma==0.1.4
langchain-cli==0.0.31
langchain-community==0.3.3
langchain-core==0.3.12
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-ollama==0.2.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langdetect==1.0.9
langserve==0.3.0
langsmith==0.1.125
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmh3==5.0.0
mock==5.1.0
monotonic==1.6
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.3.0
mteb==1.16.5
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.10.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.19.2
openai==1.52.2
openfile==0.0.7
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.11.1
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
polars==1.11.0
portalocker==2.10.1
posthog==3.6.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.2.0
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytrec-eval-terrier==0.5.6
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
ruff==0.7.1
s3transfer==0.10.3
sacrebleu==2.4.3
sacremoses==0.1.1
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.1.1
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
SpeechRecognition==3.11.0
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==1.4.54
sqlalchemy-migrate==0.11.0
sqlparse==0.5.1
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.41.0
streamlit==1.39.0
sympy==1.13.3
tabulate==0.9.0
tbb==2021.13.1
Tempita==0.5.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
toolwrapper==2.1.0
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uctools==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.23.2
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xlrd==0.7.1
xlwt==0.7.2
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
